Remove module syntax and add self references
Modules still exist in the semantics, but the syntax no longer includes
a name, using the new self references to refer to it instead. Modules
also now end in an expression, which will be executed by the dynamic
semantics. The semantic type information has been removed in favour of
just reusing the existing type syntax interpreted semantically.

Note that typing of requests currently does not work, as resolving name
bindings in other objects doesn't appear to work when the object is not
itself named.