Implement dynamic semantics
The syntax of methods has changed to merge together the statements and
result, so that the evaluation of a def can introduce their name into
the environment of the following expressions. There is a single value
sort of the reduction, a Value, which is a map from strings to methods.
The context of Self begins as an Init (program initialising, no self
value yet) and transfers to the value of the encompassing module once
that is created.