Add mutable variables
The new var syntax requires the ability to distinguish between variables
which are constant or not, so the typing of variables has been modified
to include a tag value which marks it as Def or Var. An assignment is
only well-typed if the type of the variable it is assigning to has the
Var tag.

The dynamic semantics have been modified in the same way: now the
environment values are stored inside either a Def or Var constructor,
and the arrow for Assign is only defined if the variable in the
environment is wrapped in a Var.