Adapt the language to the object inheritance spec
Top-level programs are just regular terms, with self access before
entering an object raising an exception. Type declarations have been
removed from the syntax. The syntax now distinguishes between the
presence of parameters and arguments in methods and requests, and has a
separate set of syntax for assignment methods and requests.

All typing has been removed, as it is not useful without the ability to
do structural subtyping. The dynamic semantics have been updated to the
new forms.